% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generateTweetRecommendations.R
\name{generateTweetRecommendations}
\alias{generateTweetRecommendations}
\title{generateTweetRecommendations : Given a new tweet, return n recommended tweets.}
\usage{
generateTweetRecommendations(tweetDF, tweet = NULL, num_recs = 10)
}
\arguments{
\item{tweetDF}{A data.frame returned from the prior step that called twitterBuildDatabase().}

\item{tweet}{The seed string for which recommendations will be returned for.}

\item{num_recs}{The number of recommendations requested.}
}
\value{
tweets recommended for the seed tweet.
}
\description{
\code{generateTweetRecommendations}\cr
Recommends tweets in the corpus given a seed tweet.
}
\details{
This function constructs a DocumentTermMatrics (reference: package tm) that
is weighted by TF-IDF in order to show similar words to any
given tweet. Given a seed tweet, this function adds the tweet to the
corpus, calculates TF-IDF on the entire set, and recommends n similar
tweets based on text similarity.
}
\examples{
generateTweetRecommendations(sfTweets, tweet="sourdough", num_recs=10)
}

