% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/twitterBuildDatabase.R
\name{twitterBuildDatabase}
\alias{twitterBuildDatabase}
\title{shankarz.TexST : Get tweets from Twitter, given user credentials and tokens.}
\usage{
twitterBuildDatabase(search.string = "", lat = "", long = "",
  miles = "", since = "", numberOfTweets = 1)
}
\arguments{
\item{search.string}{An optional search string to query Twitter.}

\item{lat}{A string of latitude numerics.}

\item{long}{A string of longitude numeric.}

\item{miles}{The number of miles for radius of search (as string).}

\item{since}{The start date for the Twitter search ("yyyy-mm-dd").}

\item{numberOfTweets}{Number of tweets requested from Twitter.}
}
\value{
A data.frame of twitter tweets.
}
\description{
\code{twitterBuildDatabase} \cr
Gets tweets from within an area specified by within-miles of a location specified by lattitude and longitude
}
\details{
This function conducts a search that pulls tweets that originated within miles
of lat, long that have been tweeted after the date specified by the since parameter.
Optionally, this function also accepts a search string for the query.
}
\examples{
message("To run twitterBuildDatabase, it is mandatory to have have run TwitterAuthenticate()")
 \dontrun{
 # This routine will work but because it needs internet access, it is disabled
 # for example check. This will ensure no unexpected failure during checks performed
 # during verification of the Stat290 final project
 twitterAuthenticate()

 twitterBuildDatabase(search.string = "Giants", lat="37.794108", long="-122.39511",
           miles="15", since="2014-10-01", numberOfTweets = 100)
 }
}

